requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
et-xmlfile==1.1.0
idna==3.4
numpy==1.23.4
openai==0.24.0
openpyxl==3.0.10
pandas==1.5.1
pandas-stubs==1.5.1.221024
python-dateutil==2.8.2
python-dotenv==0.21.0
pytube==12.1.0
pytz==2022.6
requests==2.28.1
six==1.16.0
tqdm==4.64.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.12
